Refactor: Move account-related classes from user package to account package
