+++ src/main/java/cz/kocabek/animerecomedationsystem/account/repository/AccWatchlistRepository.java
@@ -17,7 +17,7 @@ import cz.kocabek.animerecomedationsystem.account.entity.AccWatchlistId;
 public interface AccWatchlistRepository extends JpaRepository<AccWatchlist, AccWatchlistId> {
 
     @Query("""
-            select new cz.kocabek.animerecomedationsystem.user.dto.AccWatchlistDto( a.id.animeId,a.inWatchlist)
+            select new cz.kocabek.animerecomedationsystem.account.dto.AccWatchlistDto( a.id.animeId,a.inWatchlist)
                         from AccWatchlist a where a.id.accId=?1
             """
     )
@@ -34,9 +34,9 @@ public interface AccWatchlistRepository extends JpaRepository<AccWatchlist, AccW
 
     /*TODO choose fields and create dto for it
 
-    */
+     */
     @Query("""
-            select new cz.kocabek.animerecomedationsystem.user.dto.AccWatchlistShowDto(a.id.animeId,a.anime.name,a.anime.englishName,a.anime.score,a.anime.imageURL)
+            select new cz.kocabek.animerecomedationsystem.account.dto.AccWatchlistShowDto(a.id.animeId,a.anime.name,a.anime.englishName,a.anime.score,a.anime.imageURL)
                         from AccWatchlist a
                         where a.id.accId =:accId
                         and a.inWatchlist=true
